# Optimize has_dynamic_partition