## Summary & Motivation
The present `has_dynamic_partition` implementation is linear time, which in some cases renders it intolerably slow quite soon. For example, the `image_sensor`-scheme in the [docs](https://docs.dagster.io/concepts/partitions-schedules-sensors/partitions) starts taking several seconds at a few thousand partitions on a typical development machine.

This change moves the check from Python into the SQLAlchemy query, making it much faster. `DynamicPartitionsTable` is already conveniently indexed.

## How I Tested These Changes

Before:
```
>>> instance = DagsterInstance.ephemeral()
... instance.add_dynamic_partitions("def name", [str(uuid4()) for _ in range(2500)])
... %timeit -n 100 -r 1 -o instance.has_dynamic_partition("def name", "key")
...
<TimeitResult : 1.73 ms ± 0 ns per loop (mean ± std. dev. of 1 run, 100 loops each)>

>>> instance = DagsterInstance.ephemeral()
... instance.add_dynamic_partitions("def name", [str(uuid4()) for _ in range(5000)])
... %timeit -n 100 -r 1 -o instance.has_dynamic_partition("def name", "key")
...
<TimeitResult : 3.34 ms ± 0 ns per loop (mean ± std. dev. of 1 run, 100 loops each)>

>>> instance = DagsterInstance.ephemeral()
... instance.add_dynamic_partitions("def name", [str(uuid4()) for _ in range(10000)])
... %timeit -n 100 -r 1 -o instance.has_dynamic_partition("def name", "key")
...
<TimeitResult : 7.06 ms ± 0 ns per loop (mean ± std. dev. of 1 run, 100 loops each)>
```

After:
```
>>> instance = DagsterInstance.ephemeral()
... instance.add_dynamic_partitions("def name", [str(uuid4()) for _ in range(100000)])
... %timeit -n 100 -r 1 -o instance.has_dynamic_partition("def name", "key")
...
<TimeitResult : 402 µs ± 0 ns per loop (mean ± std. dev. of 1 run, 100 loops each)>
```